Refactor SearchFilter component
In the end I decided to split Search Filter between two
components depending on the size of the viewport. It didn't
make sense for the Sidebar component to contain that much
media query logic. Instead Category container will run the
media query and display either Modal or Sidebar version of
the component appropriately.