Change ProductSizeSelector styling
Made the cursor display as pointer when hover over the selection labels.
Also changed the font on currently selected label to regular instead of
medium weight. Having a label switch to bolder text when it becomes
selected had the side effect of making it wider, causing a weird layout
shift between the different buttons.