Remove the label from inside SpinButton
The original idea was to include the label element inside the component
to encourage the use of the label for accessibility purposes.
Unfortunately that makes positioning the label limiting, so I decided
to remove it from the component and instead it has to be added manually
every time the component is used.