Fix issue with removing characters in SpinButton
As a result of the controlled Quantity input field having '1' as minimum
value, the user wasn't able to empty the field, since that would send an
empty string, which gets converted to 0 as a number. Adding local state
to SpinButton lets you work around, while maintaining value restrictions
in the main Quantity state.