Add SignupForm component
This component will eventually fetch zxcvbn from the CDN and add it to
the DOM using useScript hook. I followed the zxcvbn developer's advice
and decided to not include it as a permanent part of the bundle, since
even minimized it's big in size and it's only used during sign up
and password reset.