Remove CartQuantityCounter component
Originally the component was intended as a client component that would
fit into ShoppingCartModal which was a server component. However with
how much interactivity the modal has, it makes no sense to split it into
a tiny server component with a bunch of client components nested inside
it. Therefore I decided to turn the whole modal into a client component
and so CartQuantityCounter is no longer necessary.