Add the generic [slug] page
This page will be used as a router of sorts, checking the 'slug' url
param against a list of preset values and displaying the appropriate
page.

This approach isn't as clean as putting each page behind an appropriate
slug (e.g. /products/ or /category/), but it allows for shorter URLs
that avoid redundant segments (i.e. elevadesign.com/furniture instead of
elevadesign.com/category/furntiure), which is better for user experience
and SEO.