Update eslint rules
Removed the rule that forbids prefixing variables with underscores.

Mongo stores its ID's in a property called _id, so eslint was showing
an error every time the property was accessed. It could technically be
worked around by changing the return documents in an aggregation stage
during database fetches, but I think it's more convenient to just
disable the rule considering it's not particularly helpful regardless.